added auth by gates for edit and delete

+++ app/Http/Controllers/QuestionsController.php
@@ -69,8 +69,15 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        //
-        return view("questions.edit", compact('question'));
+        // if(\Gate::allows('update-question', $question)) {
+        //     return view("questions.edit", compact('question'));
+        // }
+        // abort(403, "Access denied");
+
+        if(\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
+        return view('questions.edit', compact('question'));
     }
 
     /**
@@ -82,7 +89,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
-        //
+        if(\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
+
         $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', "Your question has been updated");
     }
@@ -95,7 +105,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
-        //
+        if(\Gate::denies('delete-question', $question)) {
+            abort(403, "Access denied");
+        }
+
         $question->delete();
         return redirect('/questions')->with('success', "Your question has been deleted");
     }
